Make Anagram - Alice is taking a cryptography class and finding anagrams to be very useful. We consider two strings to be anagrams of each other if the first string's letters can be rearranged to form the second string. In other words, both strings must contain the same exact letters in the same exact frequency For example, bacdc and dcbac are anagrams, but bacdc and dcbad are not.
Alice decides on an encryption scheme involving two large strings where
encryption is dependent on the minimum number of character deletions
required to make the two strings anagrams. Can you help her find this
number?

Given two strings, a and b, that may or may not be of the same length,
determine the minimum number of character deletions required to make a
and b anagrams. Any characters can be deleted from either of the
strings.

Sherlock - Watson gives Sherlock an array A of length N. Then he asks
him to determine if there exists an element in the array such that the
sum of the elements on its left is equal to the sum of the elements on
its right. If there are no elements to the left/right, then the sum is
considered to be zero.
Formally, find an i, such that A1, A2, ...Ai-1 = Ai+1, Ai+2, ..AN.